restructure code/created avatar route, model & controller
