+++ controllers/avatarController.js
@@ -1,0 +1,157 @@
+const response = require("../middleware/response");
+const User = require("../models/User");
+const Avatar = require("../models/Avatar");
+const crypto = require("crypto");
+const sharp = require("sharp");
+const dotenv = require("dotenv");
+
+dotenv.config();
+
+const randImgName = (bytes = 32) => crypto.randomBytes(bytes).toString("hex");
+
+// s3 setup
+const {
+  S3Client,
+  PutObjectCommand,
+  GetObjectCommand,
+  DeleteObjectCommand,
+} = require("@aws-sdk/client-s3");
+const { getSignedUrl } = require("@aws-sdk/s3-request-presigner");
+
+const bucketName = process.env.BUCKET_NAME;
+const bucketRegion = process.env.BUCKET_REGION;
+const accessKey = process.env.ACCESS_KEY;
+const secretAccessKey = process.env.SECRET_ACCESS_KEY;
+
+const s3 = new S3Client({
+  credentials: {
+    accessKeyId: accessKey,
+    secretAccessKey: secretAccessKey,
+  },
+  region: bucketRegion,
+});
+
+class AvatarController {
+  // upload avatar
+  static async uploadAvatar(req, res) {
+    try {
+      const buffer = req.file.buffer;
+      const id = req.user.id;
+
+      // find the id in database
+      const user = await User.findOne({
+        where: {
+          id,
+        },
+      });
+
+      // // if id do not exist print error message
+      if (!user)
+        return res
+          .status(500)
+          .send(response(" User with the given ID does not exists", {}, false));
+
+      const buffer_ = await sharp(buffer)
+        .resize({ height: 500, width: 500 })
+        .toBuffer();
+
+      const params = {
+        Bucket: bucketName,
+        Key: randImgName(),
+        Body: buffer_,
+        ContentType: req.file.mimetype,
+      };
+
+      const command = new PutObjectCommand(params);
+
+      await s3.send(command, (err, data) => {
+        console.log(err);
+      });
+
+      // save image Id
+      await Avatar.create({
+        avatar: params.Key,
+        foreign_key: user.id,
+      });
+
+      res.send("uploaded successfully");
+    } catch (error) {
+      res.status(404).send();
+    }
+  }
+
+  // fetch avatar
+  static async fetchAvatar(req, res) {
+    try {
+      const id = req.user.id;
+
+      const user = await User.findOne({
+        attributes: { exclude: ["password"] },
+        where: { id },
+      });
+
+      if (!user) {
+        return res.status(404).send(response("Faild to fetch user", {}, false));
+      }
+
+      if (user.accountStatus === "notActive") {
+        return res.status(404).send(response("Invalid Credentials", {}, false));
+      }
+
+      const getObjectParams = {
+        Bucket: bucketName,
+        Key: user.passport,
+      };
+
+      const command = new GetObjectCommand(getObjectParams);
+      const url = await getSignedUrl(s3, command, { expiresIn: 60 });
+
+      res.send(url);
+    } catch (error) {
+      res.status(404).send();
+    }
+  }
+
+  // delete avatar
+  static async deleteAvatar(req, res) {
+    try {
+      const params = req.params.id;
+      const id = req.user.id;
+
+      // find the id in database
+      const user = await User.findOne({
+        where: {
+          id,
+        },
+      });
+
+      // if id do not exist print error message
+      if (!user)
+        return res
+          .status(500)
+          .send(response(" User with the given ID does not exists", {}, false));
+
+      const getObjectParams = {
+        Bucket: bucketName,
+        Key: params,
+      };
+
+      // delete image from s3 bucket
+      const commmand = new DeleteObjectCommand(getObjectParams);
+
+      // remove image Id from db
+      await User.update(
+        {
+          passport: "",
+        },
+        { where: { id: id } }
+      );
+
+      res.send({});
+    } catch (error) {
+      res.status(404).send();
+    }
+  }
+}
+
+module.exports = AvatarController;
