+++ routes/auth.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const { body, validationResult } = require("express-validator");
+const { body } = require("express-validator");
 const { authUser, isAdmin } = require("../middleware/jwt");
 const {
   listUsers,
@@ -8,26 +8,7 @@ const {
   loginUser,
   deleteUserById,
   getUser,
-  uploadUserPassport,
-  deleteUserPassport,
-  fetchUserPassport,
 } = require("../controllers/userController");
-
-const multer = require("multer");
-const storage = multer.memoryStorage();
-const upload = multer({
-  limits: {
-    fileSize: 1000000,
-  },
-  fileFilter(req, file, cb) {
-    if (!file.originalname.match(/\.(jpg|png|jpeg)$/)) {
-      return cb(new Error("Please upload a jpg or png"));
-    }
-
-    cb(undefined, true);
-  },
-  storage: storage,
-});
 
 // @route GET api/auth
 // @desc  Get logged in User // fetch user by id*
@@ -55,29 +36,6 @@ router.post(
 // @desc  Update user*
 // @access Private
 router.put("/", [authUser], updateUserById);
-
-// @route POST api/auth/passport
-// @desc  upload passport
-// @access Private
-router.post(
-  "/passport",
-  [authUser],
-  upload.single("upload"),
-  uploadUserPassport,
-  (error, req, res, next) => {
-    res.status(400).send({ error: error.message });
-  }
-);
-
-// @route DELETE api/auth/passport
-// @desc  delete passport
-// @access Private
-router.delete("/passport", [authUser], deleteUserPassport);
-
-// @route FETCH api/auth/passport
-// @desc  get passport
-// @access Private
-router.get("/passport", [authUser], fetchUserPassport);
 
 // @route DELETE api/auth
 // @desc  Delete user account
